pyproject.toml
[project]
name = "morning-agents"
version = "0.1.0"
description = "Morning briefing agents for dev environment health, PRs, and day ahead"
requires-python = ">=3.13"
dependencies = [
    "anthropic",
    "mcp",
    "pydantic",
    "rich",
    "typer",
    "semver",
    "pyjwt>=2.12.0",
    "pyyaml",
]

[project.scripts]
morning-agents = "morning_agents.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["morning_agents"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["evals"]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
]
